Refactor
Move recovery.bin.go functions to other files and break down main() function. Add some more tests also.
