+++ .gitignore
@@ -1,1 +1,3 @@
 *.img
+
+.vscode
